Extract utility functions and GitHub helpers from release-runtime.ts (#732)
